Added a functional order page and documentation

+++ website/orders/views.py
@@ -5,22 +5,38 @@ from .forms import ProductForm, OrderRemoveForm
 
 
 class ShiftView(TemplateView):
+    """View for displaying all active shifts."""
 
     template_name = "orders/shifts.html"
 
     def get(self, request, **kwargs):
+        """
+        GET request for Shift view.
 
+        :param request: the request
+        :param kwargs: keyword arguments
+        :return: the shift view with all active shifts
+        """
         active_shifts = [x for x in Shift.objects.all() if x.is_active]
 
         return render(request, self.template_name, {"shifts": active_shifts})
 
 
 class OrderView(TemplateView):
+    """View for displaying items that can be ordered and already ordered items."""
 
     template_name = "orders/order.html"
 
     @staticmethod
     def add_order(shift, product_id, user):
+        """
+        Add an order for a specific product id, shift and user.
+
+        :param shift: the shift to add the order to
+        :param product_id: the product id of the product that the order must have
+        :param user: the user that ordered the product
+        :return: a string with an error message if the order could not be placed, True otherwise
+        """
         try:
             product = Product.objects.get(pk=product_id)
         except Product.DoesNotExist:
@@ -39,15 +55,33 @@ class OrderView(TemplateView):
 
     @staticmethod
     def delete_order(shift, order_id, user):
+        """
+        Delete an order for a specific product id, shift and user.
+
+        :param shift: the shift of the order
+        :param order_id: the id of the order to delete
+        :param user: the user that ordered the order
+        :return: a string with an error message if the order could not be deleted, True otherwise
+        """
         try:
             order = Order.objects.get(pk=order_id, user=user, shift=shift)
         except Order.DoesNotExist:
             return "That order does not exist"
+        if not shift.can_order:
+            return "You can't delete orders for this shift"
         order.delete()
         return True
 
     @staticmethod
     def get_orders_with_forms(user, shift):
+        """
+        Get a QuerySet of the orders with the .form parameter set to an OrderRemoveForm.
+
+        :param user: the user to get the orders for
+        :param shift: the shift to get the orders for
+        :return: a QuerySet of orders with an extra variable per order .form that contains an OrderRemoveForm for that
+        order
+        """
         ordered_items = Order.objects.filter(user=user, shift=shift)
         for order in ordered_items:
             order.form = OrderRemoveForm(order=order)
@@ -55,49 +89,68 @@ class OrderView(TemplateView):
 
     @staticmethod
     def get_products_with_forms(shift):
+        """
+        Get a QuerySet of products with the .form parameter set to a ProductForm.
+
+        :param shift: the shift to get the products for
+        :return: a QuerySet of products with an extra variable per product .form that contains a ProductForm for that
+        product
+        """
         items = Product.objects.filter(available=True, available_at=shift.venue)
         for item in items:
             item.form = ProductForm(product=item)
         return items
 
     def get(self, request, **kwargs):
-        shift = kwargs.get('shift')
+        """
+        GET request for Order view.
+
+        :param request: the request
+        :param kwargs: keyword arguments
+        :return: the product view with all ordered items of the user and all items that the user can order for this
+        shift
+        """
+        shift = kwargs.get("shift")
         ordered_items = self.get_orders_with_forms(request.user, shift)
         items = self.get_products_with_forms(shift)
 
-        parameters = {
-            "shift": shift,
-            "items": items,
-            "ordered_items": ordered_items
-        }
+        parameters = {"shift": shift, "items": items, "ordered_items": ordered_items}
 
         return render(request, self.template_name, parameters)
 
     def post(self, request, **kwargs):
-        shift = kwargs.get('shift')
+        """
+        POST request for Order view.
+
+        :param request: the request
+        :param kwargs: keyword arguments
+        :return: the product view with all ordered items of the user, all items that the user can order for this shift
+        and an error message if an error occurred while processing a form request for the user
+        """
+        shift = kwargs.get("shift")
         ordered_items = self.get_orders_with_forms(request.user, shift)
         items = self.get_products_with_forms(shift)
 
-        parameters = {
-            "shift": shift,
-            "items": items,
-            "ordered_items": ordered_items
-        }
+        parameters = {"shift": shift, "items": items, "ordered_items": ordered_items}
 
-        product_id = request.POST.get('product_id', None)
-        order_id = request.POST.get('order_id', None)
+        product_id = request.POST.get("product_id", None)
+        order_id = request.POST.get("order_id", None)
 
         if product_id is not None:
             error = self.add_order(shift, product_id, request.user)
             if isinstance(error, str):
-                parameters['error'] = error
+                parameters["error"] = error
             else:
-                parameters['ordered_items'] = self.get_orders_with_forms(request.user, shift)
+                parameters["ordered_items"] = self.get_orders_with_forms(
+                    request.user, shift
+                )
         elif order_id is not None:
             error = self.delete_order(shift, order_id, request.user)
             if isinstance(error, str):
-                parameters['error'] = error
+                parameters["error"] = error
             else:
-                parameters['ordered_items'] = self.get_orders_with_forms(request.user, shift)
+                parameters["ordered_items"] = self.get_orders_with_forms(
+                    request.user, shift
+                )
 
         return render(request, self.template_name, parameters)
